+++ main.rb
@@ -1,10 +1,11 @@
-require './game.rb'
+require_relative 'game'
+require_relative 'board'
+require_relative 'player'
+require_relative 'game_cycle'
+include Game_cycle
 
-game = Game.new()
-game.switch_player
-game.populate_display
-until game.over?
-  game.switch_player
-  game.play_turn
-  game.populate_display
-end
+board = Board.new
+player1 = Player.new('Player_1', x = nil,board )
+player2 = Player.new('Player_2',x = player1.token,board)
+game_object = Game.new(player1,player2,board)
+game_cycle(game_object)
