+++ spec/game_spec.rb
@@ -1,0 +1,8 @@
+# require './game'
+
+# describe '# initialize game class' do
+#     @game = Game.new
+
+
+    
+# end
